Remove effects active flag in addressable light
This was necessary to prevent overwriting the effect in
AddressableLight.update_state(), which used to be write_state() and
called in every loop iteration. However, update_state() is now only
called when the light state changes, before the effect is started, so we
don't have to worry about overwriting the effect.