Introduce AddressableLightValues API for accessing addressable pixels
This layer sits between the AddressableLight component, and the
underlying array with pixel values. It must be implemented for each
light component separately, and provides the low-level API to access
the pixel data. This API has more flexibility and can potentially
be implemented with higher performance than the API it replaces,
AddressableLight.get_view_internal().