Move saving of light values to a method on LightState
This is in preparation of splitting LightCall and LightState into
separate files, but we want to keep struct LightStateRTCState to be
LightState-private and not exposed in a header file.

The change from saving `v` to `remote_values` is safe, as before saving
the values from `v` are always written to `remove_values`.