Fix brightness increase during turn-off transition
When turning the light off, currently both the state and brightness are
independently linearly interpolated, and the product of them is used as
the output. However, the brightness of a turned-off light is set to one
in LightColorValues. For lights that initially have a low brightness,
this causes them to actually increase in brightness before turning off.

There was already a check to prevent the inverse situation in the
transition transformer, so add a check there.